Delete deprecated commands
- move git command to expo init
