+++ packages/expo-cli/src/commands/init.ts
@@ -1,5 +1,6 @@
 import { AndroidConfig, BareAppConfig, ExpoConfig, IOSConfig, getConfig } from '@expo/config';
 import * as PackageManager from '@expo/package-manager';
+import spawnAsync from '@expo/spawn-async';
 import { Exp, IosPlist, UserManager } from '@expo/xdl';
 import chalk from 'chalk';
 import program, { Command } from 'commander';
@@ -273,9 +274,45 @@ async function action(projectDir: string, command: Command) {
   try {
     // check if git is installed
     // check if inside git repo
-    await Exp.initGitRepoAsync(projectPath, { silent: true, commit: true });
+    await initGitRepoAsync(projectPath, { silent: true, commit: true });
   } catch {
     // todo: check if git is installed, bail out
+  }
+}
+
+export async function initGitRepoAsync(
+  root: string,
+  flags: { silent: boolean; commit: boolean } = { silent: false, commit: true }
+) {
+  // let's see if we're in a git tree
+  try {
+    await spawnAsync('git', ['rev-parse', '--is-inside-work-tree'], {
+      cwd: root,
+    });
+    !flags.silent && log('New project is already inside of a git repo, skipping git init.');
+  } catch (e) {
+    if (e.errno === 'ENOENT') {
+      !flags.silent && log.warn('Unable to initialize git repo. `git` not in PATH.');
+      return false;
+    }
+  }
+
+  // not in git tree, so let's init
+  try {
+    await spawnAsync('git', ['init'], { cwd: root });
+    !flags.silent && log('Initialized a git repository.');
+
+    if (flags.commit) {
+      await spawnAsync('git', ['add', '--all'], { cwd: root, stdio: 'ignore' });
+      await spawnAsync('git', ['commit', '-m', 'Created a new Expo app'], {
+        cwd: root,
+        stdio: 'ignore',
+      });
+    }
+    return true;
+  } catch (e) {
+    // no-op -- this is just a convenience and we don't care if it fails
+    return false;
   }
 }
